LayoutTests/fast/profiler/window-dot-eval.html

<html>
<head>
<script src="resources/profiler-test-JS-resources.js"></script>
<script>
if (window.layoutTestController)
    layoutTestController.dumpAsText();

console.profile("Call window.eval()");
function testStart()
{
    evalFunction();
    endTest();
}
</script>
<script> window.eval("function evalFunction() { insertNewText(); }");</script>
</head>

<body onload="testStart()">
This page has a call to window.eval().
<br>
<br>
To run this test manually, load it in the browser then load the WebInspector and look at
the profile.  In the profile there should be a call to eval().
<div id="output"></div>
</body>
</html>
