LayoutTests/fast/profiler/window-dot-eval-expected.txt

CONSOLE MESSAGE: line 59: Call window.eval()
CONSOLE MESSAGE: line 75: Thread_1  0
CONSOLE MESSAGE: line 75:  (program) window-dot-eval.html 8
CONSOLE MESSAGE: line 75:   (program) window-dot-eval.html 4
CONSOLE MESSAGE: line 75:   (program) window-dot-eval.html 15
CONSOLE MESSAGE: line 75:    eval  0
CONSOLE MESSAGE: line 75:     (program)  1
CONSOLE MESSAGE: line 75:   onload window-dot-eval.html 18
CONSOLE MESSAGE: line 75:    testStart window-dot-eval.html 10
CONSOLE MESSAGE: line 75:     evalFunction  1
CONSOLE MESSAGE: line 75:      insertNewText profiler-test-JS-resources.js 17
CONSOLE MESSAGE: line 75:       createElement  0
CONSOLE MESSAGE: line 75:       createTextNode  0
CONSOLE MESSAGE: line 75:       appendChild  0
CONSOLE MESSAGE: line 75:       getElementById  0
CONSOLE MESSAGE: line 75:     endTest profiler-test-JS-resources.js 1
CONSOLE MESSAGE: line 75:  (idle)  0
This page has a call to window.eval(). 

To run this test manually, load it in the browser then load the WebInspector and look at the profile. In the profile there should be a call to eval().
This is inserted Text
